+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Coming soon
-- tbd
+[0.11.0] - 2020-12-nn
+- Updates compatibility with `io.opentelemetry:opentelemetry-sdk:0.11.0`
+- Updates compatibility with `io.opentelemetry.javaagent:opentelemetry-javaagent-spi:0.11.0`
 
 ## [0.10.0] - 2020-11-13
 - Updates compatibility with `io.opentelemetry:opentelemetry-sdk:0.10.0`